requirements.txt
jupyter
numpy
scikit-learn
sentence-transformers
langchain-text-splitters
langchain-community
langgraph
langchain[openai]
langchain-core
duckduckgo-search
fastapi
uvicorn
playwright
langchain-apify
pyyaml
python-box
langchain-chroma
langchain-huggingface